// MangoTech/Hospital/src/app/user/user/model/user.model.ts
export class UserModel {
  id: string;
  username: string;
  password: string;
  fullName: string;
  lastName: string;
  surname: string;
  email: string;
  phone: string;
  birthDate: number;
  gender: number;
  modified: number;
  modifier: number;
  groupId: string;
  groupName: string;
  created: number;
  creator: string;
  status: number;
  removed: boolean;
}